Prepare script to deploy all on Bump.sh
How to use:

call the bash script with folderDate as parameter,
in order to build a proper changelog.

for example:
`bash scripts/deploy-all.sh 2024.06.1`
=> this will deploy new versions for 3 APIs on Bump.sh,
based on every OpenAPI document in filder dist/2024.06.1

About branch:
idp-minimal.yaml is deployed on slug github-idp:
on branch minimal (this is default branch)

idp-noEnums-minimal.yaml is deployed on same slug,
on branch noEnums-minimal (not default).